+++ HISTORY.md
@@ -2,6 +2,7 @@
 - Updated FST library to match GtkWave 3.3.79
 - The LXT wave output format is deprecated, use FST instead
 - Fix incorrect file name in assertion message (#387)
+- Fix crash while recovering from parse error (#388)
 
 ## 1.4 - 2018-07-16
 - Windows with MSYS2 is now fully supported